+++ src/ServiceBus/ServiceBus/ChangeLog.md
@@ -19,6 +19,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Moved cmdlets to V4.
 
 ## Version 3.1.1
 * Introduced secrets detection feature to safeguard sensitive data.